Remove admin pages
This reverts commit 0dd77e30e5d6ef399aaab157f4136db2d4cc3fe6.
