+++ lib/teiserver_web/controllers/admin/match_controller.ex
@@ -68,128 +68,12 @@ defmodule TeiserverWeb.Admin.MatchController do
 
   @spec show(Plug.Conn.t(), Map.t()) :: Plug.Conn.t()
   def show(conn, %{"id" => id}) do
-    match =
-      Battle.get_match!(id,
-        joins: [],
-        preload: [:members_and_users]
-      )
-
-    members =
-      match.members
-      |> Enum.sort_by(fn m -> m.user.name end, &<=/2)
-      |> Enum.sort_by(fn m -> m.team_id end, &<=/2)
-
-    match
-    |> MatchLib.make_favourite()
-    |> insert_recently(conn)
-
-    match_name = MatchLib.make_match_name(match)
-
-    rating_logs =
-      Game.list_rating_logs(
-        search: [
-          match_id: match.id
-        ]
-      )
-      |> Map.new(fn log -> {log.user_id, log} end)
-
-    # Creates a map where the party_id refers to an integer
-    # but only includes parties with 2 or more members
-    parties =
-      members
-      |> Enum.group_by(fn m -> m.party_id end)
-      |> Map.drop([nil])
-      |> Map.filter(fn {_id, members} -> Enum.count(members) > 1 end)
-      |> Map.keys()
-      |> Enum.zip(Teiserver.Helper.StylingHelper.bright_hex_colour_list())
-      |> Map.new()
-
-    # Now for balance related stuff
-    partied_players =
-      members
-      |> Enum.group_by(fn p -> p.party_id end, fn p -> p.user_id end)
-
-    groups =
-      partied_players
-      |> Enum.map(fn
-        # The nil group is players without a party, they need to
-        # be broken out of the party
-        {nil, player_id_list} ->
-          player_id_list
-          |> Enum.filter(fn userid -> rating_logs[userid] != nil end)
-          |> Enum.map(fn userid ->
-            %{userid => rating_logs[userid].value}
-          end)
-
-        {_party_id, player_id_list} ->
-          player_id_list
-          |> Enum.filter(fn userid -> rating_logs[userid] != nil end)
-          |> Map.new(fn userid ->
-            {userid, rating_logs[userid].value}
-          end)
-      end)
-      |> List.flatten()
-
-    past_balance =
-      BalanceLib.create_balance(groups, match.team_count,
-        algorithm: get_analysis_balance_algorithm()
-      )
-      |> Map.put(:balance_mode, :grouped)
-
-    # What about new balance?
-    new_balance = generate_new_balance_data(match)
-
-    raw_events =
-      Telemetry.list_simple_match_events(where: [match_id: match.id], preload: [:event_types])
-
-    events_by_type =
-      raw_events
-      |> Enum.group_by(
-        fn e ->
-          e.event_type.name
-        end,
-        fn _ ->
-          1
-        end
-      )
-      |> Enum.map(fn {name, vs} ->
-        {name, Enum.count(vs)}
-      end)
-      |> Enum.sort_by(fn v -> v end, &<=/2)
-
-    team_lookup =
-      members
-      |> Map.new(fn m ->
-        {m.user_id, m.team_id}
-      end)
-
-    events_by_team_and_type =
-      raw_events
-      |> Enum.group_by(
-        fn e ->
-          {team_lookup[e.user_id] || -1, e.event_type.name}
-        end,
-        fn _ ->
-          1
-        end
-      )
-      |> Enum.map(fn {key, vs} ->
-        {key, Enum.count(vs)}
-      end)
-      |> Enum.sort_by(fn v -> v end, &<=/2)
-
     conn
-    |> assign(:match, match)
-    |> assign(:match_name, match_name)
-    |> assign(:members, members)
-    |> assign(:rating_logs, rating_logs)
-    |> assign(:parties, parties)
-    |> assign(:past_balance, past_balance)
-    |> assign(:new_balance, new_balance)
-    |> assign(:events_by_type, events_by_type)
-    |> assign(:events_by_team_and_type, events_by_team_and_type)
-    |> add_breadcrumb(name: "Show: #{match_name}", url: conn.request_path)
-    |> render("show.html")
+    |> put_flash(
+      :info,
+      "/teiserver/admin/matches/:match_id is deprecated in favor of /battle/:id"
+    )
+    |> redirect(to: ~p"/battle/#{id}")
   end
 
   @spec user_show(Plug.Conn.t(), Map.t()) :: Plug.Conn.t()
@@ -230,42 +114,5 @@ defmodule TeiserverWeb.Admin.MatchController do
     |> assign(:uuid, uuid)
     |> assign(:matches, matches)
     |> render("server_index.html")
-  end
-
-  defp generate_new_balance_data(match) do
-    rating_type = MatchLib.game_type(match.team_size, match.team_count)
-
-    partied_players =
-      match.members
-      |> Enum.group_by(fn p -> p.party_id end, fn p -> p.user_id end)
-
-    groups =
-      partied_players
-      |> Enum.map(fn
-        # The nil group is players without a party, they need to
-        # be broken out of the party
-        {nil, player_id_list} ->
-          player_id_list
-          |> Enum.map(fn userid ->
-            %{userid => BalanceLib.get_user_rating_rank(userid, rating_type)}
-          end)
-
-        {_party_id, player_id_list} ->
-          player_id_list
-          |> Map.new(fn userid ->
-            {userid, BalanceLib.get_user_rating_rank(userid, rating_type)}
-          end)
-      end)
-      |> List.flatten()
-
-    BalanceLib.create_balance(groups, match.team_count,
-      algorithm: get_analysis_balance_algorithm()
-    )
-    |> Map.put(:balance_mode, :grouped)
-  end
-
-  defp get_analysis_balance_algorithm() do
-    # TODO move this from config into a dropdown so it can be selected on this page
-    Application.get_env(:teiserver, Teiserver)[:analysis_balance_algorithm] || "loser_picks"
   end
 end
