--- lib/teiserver_web/templates/admin/match/show.html.heex
@@ -1,112 +1,0 @@
-<% bsname = view_colour() %>
-
-<script type="text/javascript" charset="utf-8">
-  $(function() {
-    var selected_tab = window.location.hash.substr(1).replace("_tab", "");
-    $('#the_tabs a:first').tab('show');
-    $('#the_tabs [href="#' + selected_tab + '_tab"]').tab('show');
-  });
-</script>
-
-<%= render(
-  TeiserverWeb.Admin.GeneralView,
-  "sub_menu.html",
-  Map.merge(assigns, %{active: "matches"})
-) %>
-
-<div class="row section-menu">
-  <div class="col-md-12">
-    <div class={"card border-#{bsname}"}>
-      <div class="card-body">
-        <%= render(
-          TeiserverWeb.Admin.MatchView,
-          "section_menu.html",
-          Map.merge(assigns, %{
-            show_search: false,
-            active: ""
-          })
-        ) %>
-        <hr />
-
-        <div class="float-end">
-          &nbsp;
-        </div>
-
-        <h3>
-          <%= @match_name %> &nbsp;&nbsp;&nbsp;&nbsp;
-          <%= if @match.winning_team != nil do %>
-            Team <%= @match.winning_team + 1 %> won
-            &nbsp;&nbsp;&nbsp;&nbsp;
-          <% end %>
-          Duration: <%= duration_to_str_short(@match.game_duration) %>
-        </h3>
-        <br />
-        <ul class="nav nav-tabs" role="tablist" id="the_tabs">
-          <li class="nav-item">
-            <a href="#details_tab" role="tab" class="nav-link" data-bs-toggle="tab">
-              <%= central_component("icon", icon: "fa-solid fa-file-alt") %> Details
-            </a>
-          </li>
-
-          <li class="nav-item">
-            <a href="#players_tab" role="tab" class="nav-link" data-bs-toggle="tab">
-              <%= central_component("icon", icon: "fa-solid fa-users") %> Players
-            </a>
-          </li>
-
-          <%= if @rating_logs != %{} and allow?(@conn, "Reviewer") do %>
-            <li class="nav-item">
-              <a href="#ratings_tab" role="tab" class="nav-link" data-bs-toggle="tab">
-                <%= central_component("icon", icon: Teiserver.Account.RatingLib.icon()) %> Ratings
-              </a>
-            </li>
-
-            <li class="nav-item">
-              <a href="#balance_tab" role="tab" class="nav-link" data-bs-toggle="tab">
-                <%= central_component("icon", icon: "fa-solid fa-scale-balanced") %> Balance
-              </a>
-            </li>
-          <% end %>
-
-          <%= if allow?(@conn, "Server") do %>
-            <li class="nav-item">
-              <a href="#events_tab" role="tab" class="nav-link" data-bs-toggle="tab">
-                <Fontawesome.icon
-                  icon={Teiserver.Telemetry.SimpleMatchEventLib.icon()}
-                  style="regular"
-                /> &nbsp;
-                Events
-              </a>
-            </li>
-          <% end %>
-        </ul>
-
-        <div class="tab-content">
-          <div class="tab-pane" id="details_tab" style="padding:5px;">
-            <%= render("tab_details.html", assigns) %>
-          </div>
-
-          <div class="tab-pane" id="players_tab" style="padding:5px;">
-            <%= render("tab_players.html", assigns) %>
-          </div>
-
-          <%= if @rating_logs != %{} and allow?(@conn, "Reviewer") do %>
-            <div class="tab-pane" id="ratings_tab" style="padding:5px;">
-              <%= render("tab_ratings.html", assigns) %>
-            </div>
-
-            <div class="tab-pane" id="balance_tab" style="padding:5px;">
-              <%= render("tab_balance.html", assigns) %>
-            </div>
-          <% end %>
-
-          <%= if @rating_logs != %{} and allow?(@conn, "Reviewer") do %>
-            <div class="tab-pane" id="events_tab" style="padding:5px;">
-              <%= render("tab_events.html", assigns) %>
-            </div>
-          <% end %>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
