--- lib/teiserver_web/templates/admin/match/tab_balance.html.heex
@@ -1,41 +1,0 @@
-<h4>Based on data at the time</h4>
-<table class="table">
-  <tbody>
-    <tr>
-      <td>Team 1</td>
-      <td><%= @past_balance.ratings[1] |> round(2) %></td>
-    </tr>
-    <tr>
-      <td>Team 2</td>
-      <td><%= @past_balance.ratings[2] |> round(2) %></td>
-    </tr>
-    <tr>
-      <td>Deviation</td>
-      <td><%= @past_balance.deviation %></td>
-    </tr>
-  </tbody>
-</table>
-
-<textarea name="" id="" rows={Enum.count(@past_balance.logs)} class="form-control"><%= @past_balance.logs |> Enum.join("\n") %></textarea>
-
-<hr />
-
-<h4>If balance we made using current ratings</h4>
-<table class="table">
-  <tbody>
-    <tr>
-      <td>Team 1</td>
-      <td><%= @new_balance.ratings[1] |> round(2) %></td>
-    </tr>
-    <tr>
-      <td>Team 2</td>
-      <td><%= @new_balance.ratings[2] |> round(2) %></td>
-    </tr>
-    <tr>
-      <td>Deviation</td>
-      <td><%= @new_balance.deviation %></td>
-    </tr>
-  </tbody>
-</table>
-
-<textarea name="" id="" rows={Enum.count(@new_balance.logs)} class="form-control"><%= @new_balance.logs |> Enum.join("\n") %></textarea>
