--- lib/teiserver_web/templates/admin/match/tab_details.html.heex
@@ -1,37 +1,0 @@
-<%= central_component("detail_line",
-  label: "Team count",
-  value: @match.team_count
-) %>
-
-<%= central_component("detail_line",
-  label: "Team size",
-  value: @match.team_size
-) %>
-
-<hr />
-
-<%= central_component("detail_line",
-  label: "Started",
-  value: date_to_str(@match.started, format: :ymd_hms, tz: @tz)
-) %>
-
-<%= central_component("detail_line",
-  label: "Finished",
-  value: date_to_str(@match.finished, format: :ymd_hms, tz: @tz)
-) %>
-
-<%= if allow?(@conn, "admin.dev") do %>
-  <%= central_component("detail_line",
-    label: "Tag key count",
-    value: Map.keys(@match.tags) |> Enum.count()
-  ) %>
-<% end %>
-
-<%= if allow?(@conn, "Moderator") do %>
-  <hr />
-  Match data <textarea class="form-control" style="font-family: monospace" rows="20"><%= Jason.encode!(@match.data, pretty: true) %></textarea> Match bots <textarea
-    class="form-control"
-    style="font-family: monospace"
-    rows="20"
-  ><%= Jason.encode!(@match.bots, pretty: true) %></textarea>
-<% end %>
