--- lib/teiserver_web/templates/admin/match/tab_events.html.heex
@@ -1,18 +1,0 @@
-<div class="row">
-  <div class="col">
-    <h4>By type</h4>
-    <.table id="by_type" rows={@events_by_type} table_class="table-sm">
-      <:col :let={{name, _}} label="Event"><%= name %></:col>
-      <:col :let={{_, count}} label="Count"><%= count %></:col>
-    </.table>
-  </div>
-
-  <div class="col">
-    <h4>By team and type</h4>
-    <.table id="by_type" rows={@events_by_team_and_type} table_class="table-sm">
-      <:col :let={{{team, _}, _}} label="Team"><%= team + 1 %></:col>
-      <:col :let={{{_, name}, _}} label="Event"><%= name %></:col>
-      <:col :let={{_, count}} label="Count"><%= count %></:col>
-    </.table>
-  </div>
-</div>
