--- lib/teiserver_web/templates/admin/match/tab_players.html.heex
@@ -1,109 +1,0 @@
-<% bsname = view_colour() %>
-
-<table class="table table-sm">
-  <thead>
-    <tr>
-      <th colspan="6">&nbsp;</th>
-      <th colspan="2" style="text-align: center; bbackground-color: #FEE;">Damage</th>
-      <th colspan="2" style="text-align: center; bbackground-color: #EFE;">Units</th>
-      <th colspan="2" style="text-align: center; bbackground-color: #EEF;">Metal</th>
-      <th colspan="2" style="text-align: center; bbackground-color: #FFE;">Energy</th>
-
-      <th colspan="2">&nbsp;</th>
-    </tr>
-
-    <tr>
-      <th colspan="4">Name & Party</th>
-      <th>Team</th>
-      <th>Play</th>
-
-      <th>Done</th>
-      <th>Taken</th>
-
-      <th>Killed</th>
-      <th>Prod</th>
-
-      <th>Prod</th>
-      <th>Used</th>
-
-      <th>Prod</th>
-      <th>Used</th>
-
-      <th colspan="1">Rating</th>
-      <th colspan="1">&nbsp;</th>
-    </tr>
-  </thead>
-  <tbody>
-    <%= for m <- @members do %>
-      <% rating = @rating_logs[m.user_id]
-      party_colour = @parties[m.party_id]
-
-      exit_status = calculate_exit_status(m.left_after, @match.game_duration)
-
-      play_percentage =
-        if exit_status != :stayed do
-          (m.left_after / @match.game_duration * 100) |> round
-        end %>
-      <tr>
-        <td style={"background-color: #{m.user.colour}; color: #FFF;"} width="22">
-          <%= if m.team_id == @match.winning_team do %>
-            <i class="fa-fw fa-solid fa-trophy"></i>
-          <% end %>
-        </td>
-        <td style={"background-color: #{m.user.colour}; color: #FFF;"} width="22">
-          <%= central_component("icon", icon: m.user.icon) %>
-        </td>
-        <td style={"background-color: #{rgba_css m.user.colour};"}>
-          <%= m.user.name %>
-        </td>
-
-        <%= if party_colour do %>
-          <td style={"background-color: #{party_colour};"} width="50">&nbsp;</td>
-        <% else %>
-          <td style={"background-color: #{rgba_css m.user.colour};"} width="50">&nbsp;</td>
-        <% end %>
-
-        <td><%= m.team_id + 1 %></td>
-        <td>
-          <%= case exit_status do %>
-            <% :stayed -> %>
-            <% :early -> %>
-              <i class="fa-fw fa-solid fa-clock text-warning"></i> &nbsp; <%= play_percentage %>%
-            <% :abandoned -> %>
-              <i class="fa-fw fa-solid fa-person-running text-danger"></i>
-              &nbsp; <%= play_percentage %>%
-            <% :noshow -> %>
-              <i class="fa-fw fa-solid fa-user-ninja text-info2"></i>
-          <% end %>
-        </td>
-
-        <td><%= normalize(m.stats["damageDealt"]) %></td>
-        <td><%= normalize(m.stats["damageReceived"]) %></td>
-
-        <td><%= normalize(m.stats["unitsKilled"]) %></td>
-        <td><%= normalize(m.stats["unitsProduced"]) %></td>
-
-        <td><%= normalize(m.stats["metalProduced"]) %></td>
-        <td><%= normalize(m.stats["metalUsed"]) %></td>
-
-        <td><%= normalize(m.stats["energyProduced"]) %></td>
-        <td><%= normalize(m.stats["energyUsed"]) %></td>
-
-        <td>
-          <%= if rating != nil do %>
-            <%= rating.value["rating_value"] |> round(2) %>
-          <% end %>
-        </td>
-
-        <td>
-          <a
-            href={Routes.ts_admin_match_path(@conn, :user_show, m.user_id)}
-            class={"btn btn-sm btn-outline-#{bsname}"}
-          >
-            Matches
-          </a>
-        </td>
-      </tr>
-    <% end %>
-  </tbody>
-</table>
